tarefa_mod29/ExemplosJDBCAlunos/src/java/main/br/com/rsousa/domain/Produto.java
package br.com.rsousa.domain;

/**
 * @author renan.sousa
 *
 */

public class Produto {
    private Integer id;

    private String nome;

    private Double valor;

    private Integer codigo;

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getNome() {
        return nome;
    }

    public void setNome(String nome) {
        this.nome = nome;
    }

    public Double getValor() {
        return valor;
    }

    public void setValor(Double valor) {
        this.valor = valor;
    }

    public int getCodigo() {
        return codigo;
    }

    public void setCodigo(Integer codigo) {
        this.codigo = codigo;
    }
}
